change buggy buffer dtype requirement in all acq controllers, debugging minutiae

+++ qcodes/instrument_drivers/AlazarTech/Rec_controller.py
@@ -162,8 +162,7 @@ class HD_Records_Controller(AcquisitionController):
         self.board_info = alazar.get_idn()
         self.buffer = np.zeros(self.samples_per_record *
                                self.records_per_buffer *
-                               self.number_of_channels,
-                               dtype=np.uint16)
+                               self.number_of_channels)
 
         integer_list = np.arange(self.samples_per_record, dtype=np.uint16)
         angle_list = (2 * np.pi * self.demodulation_frequency /
